+++ src/extension/register/diagnostics.ts
@@ -2,6 +2,7 @@ import * as vscode from "vscode";
 import { registerBehaviorsConfigsCheck } from "../providers/diagnostics/check-behaviors-configs";
 import { registerFileLinkChecks } from "../providers/diagnostics/check-linked-file-existence";
 import { registerModelTraitConfigsCheck } from "../providers/diagnostics/check-models-traits-configs";
+import { registerSelectableOptionsExistenceChecks } from "../providers/diagnostics/check-selectable-options-existence";
 import { registerThemeFileExistenceChecks } from "../providers/diagnostics/check-theme-file-existence";
 import { registerViewChecks } from "../providers/diagnostics/check-view-existence";
 
@@ -11,4 +12,5 @@ export function registerDiagnostics(context: vscode.ExtensionContext): void {
     registerThemeFileExistenceChecks(context);
     registerFileLinkChecks(context);
     registerViewChecks(context);
+    registerSelectableOptionsExistenceChecks(context);
 }